userManagement.py file
Went over with Yazan on the userManagement.py file here. I still need to create the editUser method, and I am also unsure if I need to pass in the request parameter from the views.py file for rendering. I'm going to look into that. Also need to update the parameters for the methods as Yazan told me we are no longer using a user SSN and a few other changes.

Co-Authored-By: YazanSalem <77870522+YazanSalem@users.noreply.github.com>
Co-Authored-By: Broderick Hebert <70716012+BMHebert@users.noreply.github.com>