Numerous changes
Found a typo in TAScheduler/Management folder. Renamed it to be correct. Edited UserProfile to include username field and renamed other fields to be more easy to read. Removed AssignmentManagement since that is outdated. Updated views.py to make use of the methods from UserManagement. Updated comments a bit in views.py as well. Also changed the html files to have slightly more readable fields and made the input types for a number of inputs more accurate.