Added in editUser functionality
Got edituser to display all users in the database and allow for a supervisor to edit any of them. To accomplish this editUser.html was edited heavily and a post method for editUser in views.py was added. Other changes include adding in functionality for editUser in UserManagement since that was not working and made a small change in createuser.html.